# fix causal mask in Tensor class