@chenyuxyz 2023-11-24
i think a fix should pair with a regression test, and preferably show that it matches torch. i am surprised that the existing `test_scaled_product_attention` does not break though

is arc a 4 choice multiple choices problem? 25% is not distinguishable from random then.