@tobias17 2023-11-24
> is arc a 4 choice multiple choices problem? 25% is not distinguishable from random then.
it looks to be the case

> i am surprised that the existing `test_scaled_product_attention` does not break though
the mask gets softmaxed immediately after leading towards the change being veeeery small. i also do not think many things use this atm, since if you want KV caches you need to create your own mask

> i think a fix should pair with a regression test, and preferably show that it matches torch.
yep that's what I figured, will get started on this