@tobias17 2023-11-24
When running before (left) and after (right) of this PRs changes, the cstyle program only had 1 line diff, where a const was used instead of a value:
![image](https://github.com/tinygrad/tinygrad/assets/24630044/4c6576a6-139a-4583-b0ad-269801fe4b18)

When the programs were compiled using CUDA, the difference this made became a lot more noticalbe:
![image](https://github.com/tinygrad/tinygrad/assets/24630044/d495b936-9d5e-4084-913b-7b666f3bfffb)