@tobias17 2023-11-24
Running some timing tests, sz = (100,32,64,512) x10 for 100 iterations (with 20 warmup steps)

Before
Avg: 0.7353 sec
Std: 0.003936 sec

After
Avg: 0.7331 sec
Std: 0.004401 sec