@chenyuxyz 2023-11-24
it matches torch too (converts the mask from ones to zeros and -inf https://github.com/pytorch/pytorch/blob/d30497f6b62007c9d1e3c38179528e9d25ac1292/aten/src/ATen/native/transformers/attention.cpp#L510)